split parser and runner to modules
